fixed error on release builds
