+++ src-core/src/file_management/mod.rs
@@ -28,7 +28,7 @@ type TrajFileName = String;
 mod diagnostics;
 mod formatter;
 
-pub use diagnostics::create_diagnostic_file;
+pub use diagnostics::{create_diagnostic_file, get_log_lines};
 
 async fn write_serializable<T: Serialize + Send>(contents: T, file: &Path) -> ChoreoResult<()> {
     let json = formatter::to_string_pretty(&contents)?;
