+++ src-tauri/src/api.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 
 use crate::tauri::TauriResult;
 use choreo_core::{
-    file_management::{self, create_diagnostic_file, WritingResources},
+    file_management::{self, create_diagnostic_file, get_log_lines, WritingResources},
     generation::remote::RemoteGenerationResources,
     spec::{
         project::{ProjectFile, RobotConfig},
@@ -31,7 +31,7 @@ macro_rules! debug_result (
         }
         #[cfg(not(debug_assertions))]
         {
-            return $result;
+            return $result.map_err(Into::into);
         }
     };
 );
@@ -187,54 +187,11 @@ pub fn cancel_all_remote_generators(app_handle: tauri::AppHandle) {
     remote_resources.kill_all();
 }
 
-fn log_lines() -> Vec<String> {
-    if let Some(dir) = dirs::data_local_dir().map(|d| d.join("logs")) {
-        tracing::debug!("Looking for log files in {:}", dir.display());
-        match std::fs::read_dir(dir) {
-            Ok(dir_content) => {
-                let mut log_files = dir_content
-                    .filter_map(|entry| entry.ok())
-                    .filter(|entry| entry.file_type().map(|ft| ft.is_file()).unwrap_or(false))
-                    .filter(|entry| entry.file_name().to_string_lossy().ends_with(".log"))
-                    .collect::<Vec<_>>();
-                log_files.sort_by_key(|entry| {
-                    entry
-                        .metadata()
-                        .map(|m| m.modified().unwrap_or(std::time::SystemTime::UNIX_EPOCH))
-                        .unwrap_or(std::time::SystemTime::UNIX_EPOCH)
-                });
-                let log_file = log_files.last().ok_or(ChoreoError::FileNotFound(None));
-                match log_file {
-                    Ok(log_file) => {
-                        return std::fs::read_to_string(log_file.path())
-                            .unwrap_or_else(|e| {
-                                tracing::error!("{e}");
-                                String::new()
-                            })
-                            .lines()
-                            .map(|line| format!("{:}\n", line))
-                            .collect::<Vec<String>>()
-                    }
-                    Err(e) => {
-                        tracing::error!("{e}");
-                        Vec::new()
-                    }
-                }
-            }
-            Err(e) => {
-                tracing::error!("{e}");
-                Vec::new()
-            }
-        }
-    } else {
-        Vec::new()
-    }
-}
 
 #[tauri::command]
 pub fn open_diagnostic_file(project: ProjectFile, trajs: Vec<TrajFile>) -> TauriResult<()> {
     tracing::debug!("Opening diagnostic file");
-    let log_lines = log_lines();
+    let log_lines = get_log_lines(dirs::data_local_dir().map(|d| d.join("logs")));
 
     tracing::debug!("Found {:} log lines", log_lines.len());
